requirements.txt
streamlit
sentence-transformers
transformers
torch
scikit-learn
faiss-cpu
numpy
pandas
accelerate>=0.21.0